Update textview styles for better theme compatability.
In the battery & notification lights preferences, the textview
attached to the custom color is locked to a color which is only
visible against light backgrounds. If a theme changes to a dark
theme, they are unreadable without modifying a core material
style.

Let's create a new style that points to a better material style
but also allows a themer to change it independently if needed.

Change-Id: I2aeed8ae9845ebe0671d11f17fbf0949dc909716
Ticket: CYNGNOS-1790