[SQUASEHD] OMS Updates 1/2
Expose color for external settings icons

Marhsmallow now allows for system apps to add settings.  An icon is
loaded from the external package and a tint is applied to the icon
so that it matches other icons in the dashboard.  The tint color
uses the colorAccent attribute from the currently applied style.
Themes can overlay the colorAccent but this color is also used for
other elements of the UI which may not be ideal.

This patch adds a new color external_tile_icon_tint_color which
points to the attribute ?android:attr/colorAccent when not themed.
This retains the original look but offers better flexibility for
theme designers.

- Ported up to 7.1.1

Change-Id: Ifb4a99a315271cdcf978a796e3dea0eaef8d7750

Expose dashboard category padding bottom

Change-Id: I5a0d3266bbe96dbe6a136f78ed84cb1ee0a6a19b

Expose switchbar background color

Change-Id: Idc8b09c921bb57c9464d6ea956cfdd9b2b95e6a8

Settings: Expose bluetooth pin confirm dialog text colors

Change-Id: I1e1eb2da23f2470a82574da7ef326d56f770eb95