Fix NPE in PanelFragment
When updating the panel after getting a second intent to the
activity, there is a small window of opportunity for the user
to close the activity before the animation ends, causing an
NPE when onFinish is called as the activity already closed down.
This change will check that there is an activity before trying
to finish it.

Bug: 150322041
Test:
For an easy way to reproduce this, turn the screen on and run the
following in a terminal:
 while true; do adb shell am start -a \
 android.settings.panel.action.INTERNET_CONNECTIVITY; \
 sleep 0.1; done
Then press the back button while the screen is attempting to
display the fragment. There should be no Settings crash.

Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: spkal01 <kalligeross@gmail.com>
Signed-off-by: Ayan Mukherjee <mukherjeeayan725@gmail.com>
Change-Id: I21a93da87c163a84a45f93d30332c18bfa9e5f85
Signed-off-by: Ayan Mukherjee <mukherjeeayan725@gmail.com>