Fix SmartForwardingActivity crash at screen rotation
When SmartForwardingActivity attempts to set the Toolbar to ActionBar
in #onCreate, an llegalStateException is thrown. This is because
SmartForwardingFragment sets the default action bar by window decor to
ActionBar first by updating the title in #onCreatePreferences.
This occurs in cases where SmartForwardingFragment is created in the
process of being SmartForwardingActivity created, such as when rotating
SmartForwardingActivity.

Use getActivity()#setTitle instead of getActionBar()#setTitle.

Test: manual
Bug: 314232525
Google: 2855525
Change-Id: I041d62605ba666ecce041086854e675d9b19ac45
Signed-off-by: Jis G Jacob <studiokeys@blissroms.org>